move UDAF expression building from sql/catalyst to sql/core
