+++ sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveSessionCatalog.scala
@@ -17,27 +17,20 @@
 
 package org.apache.spark.sql.hive
 
-import java.lang.reflect.InvocationTargetException
 import java.util.Locale
 
 import scala.util.{Failure, Success, Try}
 import scala.util.control.NonFatal
 
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.hive.ql.exec.{UDAF, UDF}
 import org.apache.hadoop.hive.ql.exec.{FunctionRegistry => HiveFunctionRegistry}
-import org.apache.hadoop.hive.ql.udf.generic.{AbstractGenericUDAFResolver, GenericUDF, GenericUDTF}
 
-import org.apache.spark.sql.AnalysisException
 import org.apache.spark.sql.catalyst.FunctionIdentifier
 import org.apache.spark.sql.catalyst.analysis.{FunctionRegistry, TableFunctionRegistry}
 import org.apache.spark.sql.catalyst.catalog._
 import org.apache.spark.sql.catalyst.expressions.{Cast, Expression}
 import org.apache.spark.sql.catalyst.parser.ParserInterface
-import org.apache.spark.sql.hive.HiveShim.HiveFunctionWrapper
 import org.apache.spark.sql.types.{DecimalType, DoubleType}
-import org.apache.spark.util.Utils
-
 
 private[sql] class HiveSessionCatalog(
     externalCatalogBuilder: () => ExternalCatalog,
@@ -47,86 +40,17 @@ private[sql] class HiveSessionCatalog(
     tableFunctionRegistry: TableFunctionRegistry,
     hadoopConf: Configuration,
     parser: ParserInterface,
-    functionResourceLoader: FunctionResourceLoader)
+    functionResourceLoader: FunctionResourceLoader,
+    functionExpressionBuilder: FunctionExpressionBuilder)
   extends SessionCatalog(
-      externalCatalogBuilder,
-      globalTempViewManagerBuilder,
-      functionRegistry,
-      tableFunctionRegistry,
-      hadoopConf,
-      parser,
-      functionResourceLoader) {
-
-  private def makeHiveFunctionExpression(
-      name: String,
-      clazz: Class[_],
-      input: Seq[Expression]): Expression = {
-    var udfExpr: Option[Expression] = None
-    try {
-      // When we instantiate hive UDF wrapper class, we may throw exception if the input
-      // expressions don't satisfy the hive UDF, such as type mismatch, input number
-      // mismatch, etc. Here we catch the exception and throw AnalysisException instead.
-      if (classOf[UDF].isAssignableFrom(clazz)) {
-        udfExpr = Some(HiveSimpleUDF(name, new HiveFunctionWrapper(clazz.getName), input))
-        udfExpr.get.dataType // Force it to check input data types.
-      } else if (classOf[GenericUDF].isAssignableFrom(clazz)) {
-        udfExpr = Some(HiveGenericUDF(name, new HiveFunctionWrapper(clazz.getName), input))
-        udfExpr.get.dataType // Force it to check input data types.
-      } else if (classOf[AbstractGenericUDAFResolver].isAssignableFrom(clazz)) {
-        udfExpr = Some(HiveUDAFFunction(name, new HiveFunctionWrapper(clazz.getName), input))
-        udfExpr.get.dataType // Force it to check input data types.
-      } else if (classOf[UDAF].isAssignableFrom(clazz)) {
-        udfExpr = Some(HiveUDAFFunction(
-          name,
-          new HiveFunctionWrapper(clazz.getName),
-          input,
-          isUDAFBridgeRequired = true))
-        udfExpr.get.dataType // Force it to check input data types.
-      } else if (classOf[GenericUDTF].isAssignableFrom(clazz)) {
-        udfExpr = Some(HiveGenericUDTF(name, new HiveFunctionWrapper(clazz.getName), input))
-        // Force it to check data types.
-        udfExpr.get.asInstanceOf[HiveGenericUDTF].elementSchema
-      }
-    } catch {
-      case NonFatal(exception) =>
-        val e = exception match {
-          case i: InvocationTargetException => i.getCause
-          case o => o
-        }
-        val errorMsg = s"No handler for UDF/UDAF/UDTF '${clazz.getCanonicalName}': $e"
-        val analysisException = new AnalysisException(errorMsg)
-        analysisException.setStackTrace(e.getStackTrace)
-        throw analysisException
-    }
-    udfExpr.getOrElse {
-      throw new InvalidUDFClassException(
-        s"No handler for UDF/UDAF/UDTF '${clazz.getCanonicalName}'")
-    }
-  }
-
-  /**
-   * Constructs a [[Expression]] based on the provided class that represents a function.
-   *
-   * This performs reflection to decide what type of [[Expression]] to return in the builder.
-   */
-  override def makeFunctionExpression(
-      name: String,
-      clazz: Class[_],
-      input: Seq[Expression]): Expression = {
-    // Current thread context classloader may not be the one loaded the class. Need to switch
-    // context classloader to initialize instance properly.
-    Utils.withContextClassLoader(clazz.getClassLoader) {
-      try {
-        super.makeFunctionExpression(name, clazz, input)
-      } catch {
-        // If `super.makeFunctionExpression` throw `InvalidUDFClassException`, we construct
-        // Hive UDF/UDAF/UDTF with function definition. Otherwise, we just throw it earlier.
-        case _: InvalidUDFClassException =>
-          makeHiveFunctionExpression(name, clazz, input)
-        case NonFatal(e) => throw e
-      }
-    }
-  }
+    externalCatalogBuilder,
+    globalTempViewManagerBuilder,
+    functionRegistry,
+    tableFunctionRegistry,
+    hadoopConf,
+    parser,
+    functionResourceLoader,
+    functionExpressionBuilder) {
 
   override def lookupFunction(name: FunctionIdentifier, children: Seq[Expression]): Expression = {
     try {
